+++ python/pybind11/cache/cache_py.h
@@ -40,7 +40,7 @@ py::dict py_config(ImageCache& cache);
 void py_image_cache_reserve(ImageCache& cache, uint32_t memory_capacity, py::kwargs kwargs);
 
 py::int_ py_preferred_memory_capacity(const py::object& img,
-                                      const std::optional<const std::vector<uint32_t>>& image_size,
+                                      const std::optional<const std::vector<uint64_t>>& image_size,
                                       const std::optional<const std::vector<uint32_t>>& tile_size,
                                       const std::optional<const std::vector<uint32_t>>& patch_size,
                                       uint32_t bytes_per_pixel);
